pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dicom2bids"
version = "0.1.0"
authors = [
  { name="Lee Mullins"},
]
description = "A tool to convert DICOM files to BIDS format and prepare for NIH upload"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pydicom",
    "nibabel",
    "numpy",
    "pandas",
    "pyyaml",
    "pathlib",
    "typing-extensions",
]

[project.scripts]
dicom2bids = "dicom2bids.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dicom2bids"]

[project.urls]
Homepage = "https://github.com/lemulli/dicom2bids"
Repository = "https://github.com/lemulli/dicom2bids.git"